scripts/requirements.txt
# Kreeda Hiring Bot - Python Processing Requirements
# Essential packages for current codebase

# Redis and BullMQ
redis>=4.5.0
bullmq>=1.0.0

# Configuration
python-dotenv>=1.0.0

# File Processing
pymupdf>=1.23.0
python-docx>=0.8.11

# AI and ML (Essential)
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# HTTP Requests (for API calls to backend)
requests>=2.28.0

# Text Processing
beautifulsoup4>=4.11.0

# Utilities
fuzzywuzzy>=0.18.0